chore: add icon and rename file register
