+++ app/src/main/res/layout/activity_register.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".ui.createaccount.CreateAccountActivity">
+    tools:context=".ui.register.RegisterActivity">
 
     <Button
         android:id="@+id/back_button"
@@ -51,6 +51,7 @@
         app:boxCornerRadiusTopEnd="100dp"
         app:boxCornerRadiusTopStart="100dp"
         app:errorEnabled="true"
+        app:startIconDrawable="@drawable/outline_person_24"
         app:layout_constraintTop_toBottomOf="@+id/main_title">
 
         <com.google.android.material.textfield.TextInputEditText
@@ -69,6 +70,7 @@
         app:boxCornerRadiusBottomEnd="100dp"
         app:boxCornerRadiusBottomStart="100dp"
         app:boxCornerRadiusTopEnd="100dp"
+        app:startIconDrawable="@drawable/outline_email_24"
         app:boxCornerRadiusTopStart="100dp"
         app:errorEnabled="true"
         app:layout_constraintTop_toBottomOf="@+id/til_name">
@@ -92,6 +94,7 @@
         app:boxCornerRadiusTopStart="100dp"
         app:endIconMode="password_toggle"
         app:errorEnabled="true"
+        app:startIconDrawable="@drawable/outline_lock_24"
         app:layout_constraintTop_toBottomOf="@+id/til_email"
         tools:layout_editor_absoluteX="16dp">
 
@@ -105,6 +108,7 @@
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/til_confirm_password"
+        app:startIconDrawable="@drawable/outline_lock_24"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:hint="@string/hint_confirm_password"
